Merge hotfix into master

Closes #1393, #1868, #1873, #1874, #1845, #1871, #1877, #1587, #1886, #1787, #1884, #1900, #1979, #1872, #1916, #1992, #1876, #2032, #2054, #1991, #2041, #2084, #1988, #2074, #2118, #2114, #2100, #2103, #2026, #2119, #2113, #2111, #2127, #2083, #2160, #2164, #2172, #1781, and #1875

See merge request emundus/cms/tchooz!1495

+++ components/com_fabrik/controllers/list.php
@@ -268,8 +268,16 @@ class FabrikControllerList extends BaseController
 	 */
 	public function doempty()
 	{
-		Html::validateRequest($this->taskMap);
+		Html::validateRequest($this->taskMap, true);
+
+		/** @var FabrikFEModelList $model */
 		$model = $this->getModel('list', 'FabrikFEModel');
+
+		if (!$model->canEmpty())
+		{
+			throw new Exception(Text::_('JERROR_ALERTNOAUTHOR'), 403);
+		}
+
 		$model->truncate();
 		$this->display();
 	}
